Support private repositories
