+++ module.yml
@@ -32,13 +32,17 @@ Parameters:
     Type: String
     Default: ''
   ManagedPolicyArns:
-    Description: 'Comma-delimited list of IAM managed policy ARNs to attach to the task''s IAM role'
+    Description: 'Comma-delimited list of IAM managed policy ARNs to attach to the task''s IAM role.'
     Type: String
     Default: ''
   AppImage:
-    Description: 'The Docker image to use for the app container. You can use images in the Docker Hub registry or specify other repositories (repository-url/image:tag).'
+    Description: 'The Docker image to use for the app container. You can use images in the Docker Hub registry or specify other repositories (repository-url/image:tag). If the repository is private, set AppImageSecretModule as well!'
     Type: String
-    Default: 'widdix/hello:v1'
+    Default: 'widdix/hello:v1' 
+  AppImageSecretModule:
+    Description: 'Optional stack name of secret module which contains the repository credentials for private registry authentication.'
+    Type: String
+    Default: ''
   AppEntryPoint:
     Description: 'Optional entry point for app container.'
     Type: String
@@ -108,7 +112,11 @@ Parameters:
     Type: String
     Default: ''
   SidecarImage:
-    Description: 'Optional Docker image to use for the sidecar container. You can use images in the Docker Hub registry or specify other repositories (repository-url/image:tag).'
+    Description: 'Optional Docker image to use for the sidecar container. You can use images in the Docker Hub registry or specify other repositories (repository-url/image:tag). If the repository is private, set SidecarImageSecretModule as well!'
+    Type: String
+    Default: '' 
+  SidecarImageSecretModule:
+    Description: 'Optional stack name of secret module which contains the repository credentials for private registry authentication.'
     Type: String
     Default: ''
   SidecarPort:
@@ -267,6 +275,7 @@ Conditions:
   HasManagedPolicyArns: !Not [!Equals [!Ref ManagedPolicyArns, '']]
   HasAppCommand: !Not [!Equals [!Join ['', !Ref AppCommand], '']]
   HasAppEntryPoint: !Not [!Equals [!Ref AppEntryPoint, '']]
+  HasAppImageSecretModule: !Not [!Equals [!Ref AppImageSecretModule, '']]
   HasAppEnvironment1PlainText: !And [!Not [!Equals [!Ref AppEnvironment1Key, '']], !Equals [!Ref AppEnvironment1SecretModule, '']]
   HasAppEnvironment2PlainText: !And [!Not [!Equals [!Ref AppEnvironment2Key, '']], !Equals [!Ref AppEnvironment2SecretModule, '']]
   HasAppEnvironment3PlainText: !And [!Not [!Equals [!Ref AppEnvironment3Key, '']], !Equals [!Ref AppEnvironment3SecretModule, '']]
@@ -277,6 +286,7 @@ Conditions:
   HasAppEnvironment2Secret: !And [!Not [!Equals [!Ref AppEnvironment2Key, '']], !Not [!Equals [!Ref AppEnvironment2SecretModule, '']]]
   HasAppEnvironment3Secret: !And [!Not [!Equals [!Ref AppEnvironment3Key, '']], !Not [!Equals [!Ref AppEnvironment3SecretModule, '']]]
   HasSidecarImage: !Not [!Equals [!Ref SidecarImage, '']]
+  HasSidecarImageSecretModule: !Not [!Equals [!Ref SidecarImageSecretModule, '']]
   HasSidecarEnvironment1PlainText: !And [!Not [!Equals [!Ref SidecarEnvironment1Key, '']], !Equals [!Ref SidecarEnvironment1SecretModule, '']]
   HasSidecarEnvironment2PlainText: !And [!Not [!Equals [!Ref SidecarEnvironment2Key, '']], !Equals [!Ref SidecarEnvironment2SecretModule, '']]
   HasSidecarEnvironment3PlainText: !And [!Not [!Equals [!Ref SidecarEnvironment3Key, '']], !Equals [!Ref SidecarEnvironment3SecretModule, '']]
@@ -313,6 +323,21 @@ Resources:
             - 'logs:CreateLogStream'
             - 'logs:PutLogEvents'
             Resource: !GetAtt 'LogGroup.Arn'
+          - Effect: Allow
+            Action: 'kms:Decrypt'
+            Resource: '*'
+          - !If
+            - HasAppImageSecretModule
+            - Effect: Allow
+              Action: 'secretsmanager:GetSecretValue'
+              Resource: {'Fn::ImportValue': !Sub '${AppImageSecretModule}-Arn'}
+            - !Ref 'AWS::NoValue'
+          - !If
+            - HasSidecarImageSecretModule
+            - Effect: Allow
+              Action: 'secretsmanager:GetSecretValue'
+              Resource: {'Fn::ImportValue': !Sub '${SidecarImageSecretModule}-Arn'}
+            - !Ref 'AWS::NoValue'
           - !If
             - HasAppEnvironment1Secret
             - Effect: Allow
@@ -387,6 +412,7 @@ Resources:
       ContainerDefinitions:
       - Name: app
         Image: !Ref AppImage
+        RepositoryCredentials: !If [HasAppImageSecretModule, {CredentialsParameter: {'Fn::ImportValue': !Sub '${AppImageSecretModule}-Arn'}}, !Ref 'AWS::NoValue']
         Essential: true
         Command: !If [HasAppCommand, !Ref AppCommand, !Ref 'AWS::NoValue']
         EntryPoint: !If [HasAppEntryPoint, !Ref AppEntryPoint, !Ref 'AWS::NoValue']
@@ -411,6 +437,7 @@ Resources:
         - HasSidecarImage
         - Name: sidecar
           Image: !Ref SidecarImage
+          RepositoryCredentials: !If [HasSidecarImageSecretModule, {CredentialsParameter: {'Fn::ImportValue': !Sub '${SidecarImageSecretModule}-Arn'}}, !Ref 'AWS::NoValue']
           PortMappings:
           - ContainerPort: !Ref SidecarPort
             Protocol: tcp
@@ -584,7 +611,7 @@ Outputs:
   ModuleId:
     Value: 'fargate-scheduled-task'
   ModuleVersion:
-    Value: '1.2.0'
+    Value: '1.3.0'
   StackName:
     Value: !Ref 'AWS::StackName'
   TaskRoleArn:
