+++ test/alerting.js
@@ -1,0 +1,13 @@
+const test = require('ava');
+const cfntest = require('@cfn-modules/test');
+
+test.serial('alerting', async t => {
+  const stackName = cfntest.stackName();
+  try {
+    t.log(await cfntest.createStack(`${__dirname}/alerting.yml`, stackName, {}));
+    // what could we test here?
+  } finally {
+    t.log(await cfntest.deleteStack(stackName));
+    t.pass();
+  }
+});
