+++ test/cluster.js
@@ -1,0 +1,13 @@
+const test = require('ava');
+const cfntest = require('@cfn-modules/test');
+
+test.serial('cluster', async t => {
+  const stackName = cfntest.stackName();
+  try {
+    t.log(await cfntest.createStack(`${__dirname}/cluster.yml`, stackName, {}));
+    // what could we test here?
+  } finally {
+    t.log(await cfntest.deleteStack(stackName));
+    t.pass();
+  }
+});
