+++ test/package.json
@@ -8,6 +8,7 @@
     "@cfn-modules/ecs-cluster": "1.1.0",
     "@cfn-modules/alerting": "1.2.0",
     "@cfn-modules/client-sg": "1.0.0",
+    "@cfn-modules/secret": "1.3.0",
     "@cfn-modules/fargate-scheduled-task": "file:../"
   },
   "scripts": {
