+++ test/private-dockerhub.yml
@@ -1,0 +1,32 @@
+---
+AWSTemplateFormatVersion: '2010-09-09'
+Description: 'cfn-modules test'
+Resources:
+  Vpc:
+    Type: 'AWS::CloudFormation::Stack'
+    Properties:
+      Parameters:
+        S3Endpoint: 'false' # speed up the test
+        DynamoDBEndpoint: 'false' # speed up the test
+        FlowLog: 'false' # speed up the test
+        NatGateways: 'false' # speed up the test
+      TemplateURL: './node_modules/@cfn-modules/vpc/module.yml'
+  Secret:
+    Type: 'AWS::CloudFormation::Stack'
+    Properties:
+      Parameters:
+        Arn: 'arn:aws:secretsmanager:us-east-1:237442020868:secret:dockerhub-michael-O01ljL'
+      TemplateURL: './node_modules/@cfn-modules/secret/wrapper.yml'
+  ScheduledTask:
+    Type: 'AWS::CloudFormation::Stack'
+    Properties:
+      Parameters:
+        VpcModule: !GetAtt 'Vpc.Outputs.StackName'
+        ScheduleExpression: 'rate(15 minutes)'
+        Timeout: '600'
+        AppImage: 'hellomichibye/hello:v1'
+        AppImageSecretModule: !GetAtt 'Secret.Outputs.StackName'
+        AppCommand: 'echo,test'
+      TemplateURL: './node_modules/@cfn-modules/fargate-scheduled-task/module.yml'
+
+
