--- test/test.js
@@ -1,46 +1,0 @@
-const test = require('ava');
-const cfntest = require('@cfn-modules/test');
-
-test.serial('defaults', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/defaults.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test.serial('cluster', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/cluster.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test.serial('alerting', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/alerting.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test.serial('client-sg', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/client-sg.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
